+++ .storybook/manager.js
@@ -10,4 +10,5 @@ addons.setConfig({
     brandImage: "https://eox.at/EOX_Logo.svg",
     brandTarget: "_self",
   }),
+  selectedPanel: "eox/storybook-custom-panels/description",
 });