+++ elements/storytelling/stories/markdown-init-event.js
@@ -7,8 +7,7 @@ function initEventFunc(element) {
   if (element?.tagName === "EOX-MAP") {
     element.registerProjection(
       "ESRI:53009",
-      "+proj=moll +lon_0=0 +x_0=0 +y_0=0 +a=6371000 " +
-        "+b=6371000 +units=m +no_defs",
+      "+proj=moll +lon_0=0 +x_0=0 +y_0=0 +a=6371000 +b=6371000 +units=m +no_defs",
     );
     const projection = element.getAttribute("projection");
     if (projection) element.setAttribute("projection", projection);
@@ -16,6 +15,13 @@ function initEventFunc(element) {
 }
 
 export const MarkdownInitEvent = {
+  argTypes: {
+    markdown: {
+      table: {
+        category: "properties",
+      },
+    },
+  },
   args: {
     markdown: `## Map section <!--{as="eox-map" projection="ESRI:53009" style="width: 100%; height: 500px;" config='{ "controls": { "Zoom": {}, "Attribution": {}, "FullScreen": {}, "OverviewMap": { "layers": [ { "type": "Tile", "properties": { "id": "overviewMap" }, "source": { "type": "OSM" } } ] } }, "layers": [ { "type": "Tile", "properties": { "id": "overviewMap" }, "source": { "type": "TileWMS", "url": "https://ows.mundialis.de/services/service", "params": { "LAYERS": "TOPO-WMS" } } } ], "view": { "center": [15,48], "zoom": 1 } }'}-->
 
@@ -30,12 +36,26 @@ export const MarkdownInitEvent = {
 ### <!--{ projection="EPSG:4326" animationOptions="{duration:500}" }-->
 #### EPSG:4326
 `,
+    id: "markdown-init-event",
+    init: (e) => initEventFunc(e.detail),
+    storyCodeBefore: `
+    function initEventFunc(element) {
+      if (element?.tagName === "EOX-MAP") {
+        element.registerProjection(
+          "ESRI:53009",
+          "+proj=moll +lon_0=0 +x_0=0 +y_0=0 +a=6371000 +b=6371000 +units=m +no_defs",
+        );
+        const projection = element.getAttribute("projection");
+        if (projection) element.setAttribute("projection", projection);
+      }
+    }
+    `,
   },
   render: (args) => html`
     <eox-storytelling
-      id="markdown-editor"
+      id=${args.id}
       markdown=${args.markdown}
-      @init=${({ detail }) => initEventFunc(detail)}
+      @init=${args.init}
     ></eox-storytelling>
   `,
 };